moved the handlePlaceSelection into the services folder
